requirements.txt
numpy==1.24.3
opencv_contrib_python==4.7.0.72
opencv_python==4.6.0.66
opencv_python_headless==4.7.0.72
Pillow==9.5.0

